requirements.txt
python-docx
lxml
python-dotenv
boto3
streamlit
